Added the OD files for testing
This program presents OD8 on the screen, creates an empty text file and copies it to the directory for the trial being presented. This will be replaced with Derek's camera program and eventually with Python code.
Open OD in psyexp and edit the experiment so that it points to the correct paths for the excel file and the camfile.bat
To DO:
Implement Shutters